Parser simplification (incomplete)
